Turn EventStore mix tasks into generic tasks for use with Distillery
Intentionally did not try and address using Ecto for migrations, as I
did not want to introduce the dependency at this time. This simply moves
the logic to eventstore/tasks and removes any Mix dependencies.
